refactor: move client classes to client package
move the main client classes (BlinkStick, BlinkStickPro, BlinkStickProMatrix) to a clients package

Also, move the core functions (find_first etc) to core module
