+++ src/blinkstick/__init__.py
@@ -1,12 +1,7 @@
 from importlib.metadata import version, PackageNotFoundError
 
-from blinkstick.clients.blinkstick import BlinkStick, BlinkStickPro, BlinkStickProMatrix
-from blinkstick.clients.blinkstick import (
-    find_all,
-    find_by_serial,
-    find_first,
-    get_blinkstick_package_version,
-)
+from blinkstick.clients import BlinkStick, BlinkStickPro, BlinkStickProMatrix
+from .core import find_all, find_first, find_by_serial, get_blinkstick_package_version
 from .colors import Color, ColorFormat
 from .constants import BlinkStickVariant
 from .exceptions import BlinkStickException
