+++ src/blinkstick/clients/__init__.py
@@ -1,0 +1,4 @@
+from .blinkstick import BlinkStick
+from .blinkstick_pro import BlinkStickPro, BlinkStickProMatrix
+
+__all__ = ["BlinkStick", "BlinkStickPro", "BlinkStickProMatrix"]
